رفع اشکال

+++ ApiKarbord/Models/Public/Web_AGMkz.cs
@@ -10,7 +10,7 @@
     {
         public short? MainLevel { get; set; }
 
-        public int? Tag { get; set; }
+        public byte? Tag { get; set; }
 
         public short? Level { get; set; }
